move TargetingRulesService functionality into DomainSettingsService
