+++ libs/state/src/core/state-definitions.ts
@@ -97,7 +97,6 @@ export const AUTOFILL_SETTINGS_DISK_LOCAL = new StateDefinition("autofillSetting
 });
 
 export const AUTOTYPE_SETTINGS_DISK = new StateDefinition("autotypeSettings", "disk");
-export const TARGETING_RULES_DISK = new StateDefinition("targetingRules", "disk");
 
 // Components
